requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
asgiref==3.4.1
asn1crypto==0.24.0
astroid==2.1.0
asttokens==1.1.13
async-timeout==4.0.2
asynctest==0.13.0
attrs==21.4.0
automationhat==0.2.0
beautifulsoup4==4.7.1
blinker==1.4
blinkt==0.1.2
buttonshim==0.0.2
cachetools==4.2.4
Cap1xxx==0.1.3
certifi==2018.8.24
chardet==3.0.4
charset-normalizer==2.0.7
Click==7.0
colorama==0.3.7
colorzero==1.1
cookies==2.2.1
cryptography==2.6.1
cupshelpers==1.0
Django==3.2.10
docutils==0.14
drumhat==0.1.0
entrypoints==0.3
envirophat==1.0.0
ExplorerHAT==0.4.2
Flask==1.0.2
fourletterphat==0.1.0
frozenlist==1.2.0
future==0.18.2
google-api-core==2.3.2
google-api-python-client==2.34.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
googleapis-common-protos==1.54.0
gpiozero==1.6.2
gunicorn==20.1.0
html5lib==1.0.1
httplib2==0.20.2
idna==2.6
image==1.5.33
imutils==0.5.4
isort==4.3.4
itsdangerous==0.24
jedi==0.13.2
Jinja2==2.10
keyring==17.1.1
keyrings.alt==3.1.1
lazy-object-proxy==1.3.1
line-bot-sdk==2.0.1
logilab-common==1.4.2
lxml==4.3.2
MarkupSafe==1.1.0
mccabe==0.6.1
microdotphat==0.2.1
mote==0.0.4
motephat==0.0.3
multidict==5.2.0
mypy==0.670
mypy-extensions==0.4.1
numpy==1.16.2
oauth2client==4.1.3
oauthlib==2.1.0
olefile==0.46
pantilthat==0.0.7
parso==0.3.1
pexpect==4.6.0
pgzero==1.2
phatbeat==0.1.1
pianohat==0.1.0
picamera==1.13
piglow==1.2.5
pigpio==1.78
Pillow==5.4.1
playsound==1.3.0
protobuf==3.19.1
psutil==5.5.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
pygame==1.9.4.post1
Pygments==2.3.1
PyGObject==3.30.4
pyinotify==0.9.6
PyJWT==1.7.0
pylint==2.2.2
pyngrok==5.1.0
pyOpenSSL==19.0.0
pyparsing==3.0.6
pyserial==3.4
pysmbc==1.0.15.6
python-apt==1.8.4.3
pytz==2021.3
pyxdg==0.25
PyYAML==6.0
rainbowhat==0.1.0
reportlab==3.5.13
requests==2.26.0
requests-oauthlib==1.0.0
responses==0.9.0
roman==2.0.0
RPi.GPIO==0.7.0
rsa==4.8
RTIMULib==7.2.1
scrollphat==0.0.7
scrollphathd==1.2.1
SecretStorage==2.3.1
Send2Trash==1.5.0
sense-hat==2.2.0
simplejson==3.16.0
six==1.12.0
skywriter==0.0.7
sn3218==1.2.7
soupsieve==1.8
spidev==3.5
sqlparse==0.4.2
ssh-import-id==5.7
thonny==3.3.10
touchphat==0.0.1
twython==3.7.0
typed-ast==1.3.1
typing_extensions==4.0.1
unicornhathd==0.0.4
uritemplate==4.1.1
urllib3==1.24.1
webencodings==0.5.1
Werkzeug==0.14.1
wrapt==1.10.11
yarl==1.7.2
